+++ docs/CHANGELOG.rst
@@ -20,6 +20,7 @@ Changed
 - Add ``sample_annotation`` Jinja expressions filter that accepts an annotation
   path and returns its value
 - Update field paths for sample annotation in ``geo-import`` process
+- Update species annotation in ``alignment-star`` process
 
 Fixed
 -----